requirements.txt
coloredlogs
requests
PyYAML
